Renaming the publisher class and tests
